gpu-node/worker/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
faiss-gpu>=1.7.4
decord>=0.6.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pyacoustid>=1.3.0
pydantic>=2.5.0
tqdm>=4.66.0
